+++ CotEditor/Sources/DefaultKeys.swift
@@ -108,7 +108,6 @@ extension DefaultKeys {
     static let encodingInNew = DefaultKey<UInt>("encodingInNew")
     static let saveUTF8BOM = DefaultKey<Bool>("saveUTF8BOM")
     static let referToEncodingTag = DefaultKey<Bool>("referToEncodingTag")
-    static let enableSyntaxHighlight = DefaultKey<Bool>("doSyntaxColoring")
     static let syntaxStyle = DefaultKey<String>("defaultColoringStyleName")
     
     // File Drop
@@ -176,7 +175,6 @@ extension DefaultKeys {
     static let headerFooterDateFormat = DefaultKey<String>("headerFooterDateFormat")
     static let headerFooterPathAbbreviatingWithTilde = DefaultKey<Bool>("headerFooterPathAbbreviatingWithTilde")
     static let autoCompletionDelay = DefaultKey<Double>("autoCompletionDelay")
-    static let showColoringIndicatorTextLength = DefaultKey<Int>("showColoringIndicatorTextLength")
     static let coloringRangeBufferLength = DefaultKey<Int>("coloringRangeBufferLength")
     static let largeFileAlertThreshold = DefaultKey<Int>("largeFileAlertThreshold")
     static let savesTextOrientation = DefaultKey<Bool>("savesTextOrientation")