+++ CotEditor/Sources/DefaultSettings.swift
@@ -105,7 +105,6 @@ struct DefaultSettings {
         .encodingInNew: String.Encoding.utf8.rawValue,
         .saveUTF8BOM: false,
         .referToEncodingTag: true,
-        .enableSyntaxHighlight: true,
         .syntaxStyle: "Plain Text",
         
         .fileDropArray: [
@@ -181,7 +180,6 @@ struct DefaultSettings {
         .headerFooterDateFormat: "yyyy-MM-dd HH:mm",
         .headerFooterPathAbbreviatingWithTilde: true,
         .autoCompletionDelay: 0.25,
-        .showColoringIndicatorTextLength: 75000,
         .coloringRangeBufferLength: 5000,
         .largeFileAlertThreshold: 50 * pow(1024, 2),  // 50 MB
         .savesTextOrientation: true,