+++ CotEditor/Sources/Pair.swift
@@ -8,7 +8,7 @@
 //
 //  ---------------------------------------------------------------------------
 //
-//  © 2016-2021 1024jp
+//  © 2016-2022 1024jp
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -66,7 +66,7 @@ extension Pair where T == Character {
 
 
 
-extension StringProtocol where Self.Index == String.Index {
+extension StringProtocol {
     
     /// Find the mate of a brace pair.
     ///