+++ CotEditor/Sources/ProgressViewController.swift
@@ -33,7 +33,6 @@ final class ProgressViewController: NSViewController {
     @objc private dynamic let progress: Progress
     @objc private dynamic let message: String
     private let closesAutomatically: Bool
-    private let appliesDescriptionImmediately: Bool
     
     private var progressSubscriptions: Set<AnyCancellable> = []
     private var completionSubscriptions: Set<AnyCancellable> = []
@@ -54,8 +53,7 @@ final class ProgressViewController: NSViewController {
     ///   - progress: The progress instance to indicate.
     ///   - message: The text to display as the message label of the indicator.
     ///   - closesAutomatically: Whether dismiss the view when the progress is finished.
-    ///   - appliesDescriptionImmediately: Whether throttle the update of the description field.
-    init?(coder: NSCoder, progress: Progress, message: String, closesAutomatically: Bool = true, appliesDescriptionImmediately: Bool = false) {
+    init?(coder: NSCoder, progress: Progress, message: String, closesAutomatically: Bool = true) {
         
         assert(!progress.isCancelled)
         assert(!progress.isFinished)
@@ -63,7 +61,6 @@ final class ProgressViewController: NSViewController {
         self.progress = progress
         self.message = message
         self.closesAutomatically = closesAutomatically
-        self.appliesDescriptionImmediately = appliesDescriptionImmediately
         
         super.init(coder: coder)
         
@@ -104,7 +101,7 @@ final class ProgressViewController: NSViewController {
             .store(in: &self.progressSubscriptions)
         
         self.progress.publisher(for: \.localizedDescription, options: .initial)
-            .throttle(for: self.appliesDescriptionImmediately ? 0 : 0.1, scheduler: DispatchQueue.main, latest: true)
+            .throttle(for: 0.1, scheduler: DispatchQueue.main, latest: true)
             .assign(to: \.stringValue, on: descriptionField)
             .store(in: &self.progressSubscriptions)
     }