+++ CotEditor/Sources/TextFind.swift
@@ -171,20 +171,25 @@ final class TextFind {
     /// Return the nearest match from the insertion point.
     ///
     /// - Parameters:
-    ///   - forward: Whether search forward from the insertion.
-    ///   - isWrap: Whetehr search wrap search around.
+    ///   - forward: Whether searchs forward from the insertion.
+    ///   - isWrap: Whether the search wraps  around.
+    ///   - includingCurrentSelection: Whether includes the current selection to search.
     /// - Returns:
     ///   - range: The range of matched or nil if not found.
-    ///   - count: The total number of matches in the scopes.
+    ///   - ranges: The ranges of all matches in the scopes.
     ///   - wrapped: Whether the search was wrapped to find the result.
-    func find(forward: Bool, isWrap: Bool) -> (range: NSRange?, count: Int, wrapped: Bool) {
+    func find(forward: Bool, isWrap: Bool, includingSelection: Bool = false) -> (range: NSRange?, ranges: [NSRange], wrapped: Bool) {
+        
+        assert(forward || !includingSelection)
         
         if self.inSelection {
             return self.findInSelection(forward: forward)
         }
         
         let selectedRange = self.selectedRanges.first!
-        let startLocation = forward ? selectedRange.upperBound : selectedRange.location
+        let startLocation = forward
+            ? (includingSelection ? selectedRange.lowerBound : selectedRange.upperBound)
+            : (includingSelection ? selectedRange.upperBound : selectedRange.lowerBound)
         
         var forwardMatches: [NSRange] = []  // matches after the start location
         let forwardRange = NSRange(startLocation..<(self.string as NSString).length)
@@ -214,21 +219,21 @@ final class TextFind {
             foundRange = forward ? (wrappedMatches + intersectionMatches).first : (intersectionMatches + forwardMatches).last
         }
         
-        let count = forwardMatches.count + wrappedMatches.count + intersectionMatches.count
+        let ranges = forwardMatches + intersectionMatches + wrappedMatches
         
-        return (foundRange, count, isWrapped)
+        return (foundRange, ranges, isWrapped)
     }
     
     
     /// Return a match in selection ranges.
     ///
     /// - Parameters:
-    ///   - forward: Whether search forward from the insertion.
+    ///   - forward: Whether searchs forward from the insertion.
     /// - Returns:
     ///   - range: The range of matched or nil if not found.
-    ///   - count: The total number of matches in the scopes.
+    ///   - ranges: The ranges of all matches in the scopes.
     ///   - wrapped: Whether the search was wrapped to find the result.
-    func findInSelection(forward: Bool) -> (range: NSRange?, count: Int, wrapped: Bool) {
+    func findInSelection(forward: Bool) -> (range: NSRange?, ranges: [NSRange], wrapped: Bool) {
         
         assert(self.inSelection)
         
@@ -239,7 +244,7 @@ final class TextFind {
         
         let foundRange = forward ? matches.first : matches.last
         
-        return (foundRange, matches.count, false)
+        return (foundRange, matches, false)
     }